Started work on separating abstraction and implementation
