+++ lib/src/api/events/ready_event.dart
@@ -1,0 +1,5 @@
+import '../../../restrr.dart';
+
+class ReadyEvent extends RestrrEvent {
+  const ReadyEvent({required super.api});
+}
