+++ lib/src/api/events/restrr_event.dart
@@ -1,0 +1,7 @@
+import '../../../restrr.dart';
+
+abstract class RestrrEvent {
+  final Restrr api;
+
+  const RestrrEvent({required this.api});
+}
