+++ lib/src/api/requests/restrr_errors.dart
@@ -1,0 +1,25 @@
+/// Represents a type of error that can occur during a REST request.
+enum RestrrError {
+  /* Request errors */
+  badRequest,
+
+  /* Server errors */
+
+  notFound,
+  notSignedIn,
+  invalidCredentials,
+  alreadySignedIn,
+  internalServerError,
+  serviceUnavailable,
+
+  /* Client errors */
+
+  noInternetConnection(clientError: true),
+  serverUnreachable(clientError: true),
+  invalidUri(clientError: true),
+  unknown(clientError: true);
+
+  final bool clientError;
+
+  const RestrrError({this.clientError = false});
+}
