+++ lib/src/api/server_info.dart
@@ -1,0 +1,5 @@
+abstract class ServerInfo {
+  bool get healthy;
+  int get apiVersion;
+  String? get details;
+}
