--- lib/src/events/event_handler.dart
@@ -1,15 +1,0 @@
-import 'package:restrr/src/events/restrr_event.dart';
-
-class RestrrEventHandler {
-  final Map<Type, Function> _eventMap;
-
-  const RestrrEventHandler(this._eventMap);
-
-  void on<T extends RestrrEvent>(Type type, Function(T) callback) {
-    _eventMap[type] = callback;
-  }
-
-  void fire<T extends RestrrEvent>(T event) {
-    _eventMap[event.runtimeType]?.call(event);
-  }
-}
