--- lib/src/events/ready_event.dart
@@ -1,5 +1,0 @@
-import 'package:restrr/src/events/restrr_event.dart';
-
-class ReadyEvent extends RestrrEvent {
-  const ReadyEvent({required super.api});
-}
