--- lib/src/events/restrr_event.dart
@@ -1,7 +1,0 @@
-import '../../restrr.dart';
-
-abstract class RestrrEvent {
-  final Restrr api;
-
-  const RestrrEvent({required this.api});
-}
