+++ lib/src/internal/utils/io_utils.dart
