+++ lib/src/internal/utils/string_utils.dart
